requirements.txt

streamlit
pandas
plotly
fpdf
pdfplumber
openpyxl
xlsxwriter
pyyaml
streamlit-authenticator
